small updates

+++ src/main/java/alisgroup/titanicmanipulator/Matrix.java
@@ -5,38 +5,38 @@ import java.lang.reflect.Method;
 import java.util.List;
 
 public class Matrix {
-    int matrixX[][];
-    long matrixY[];
+    float matrixX[][];
+    float matrixY[];
 
     public Matrix (){}
-    public Matrix(int[][] matrixX, long[] matrixY) {
+    public Matrix(float[][] matrixX, float[] matrixY) {
         this.matrixX = matrixX;
         this.matrixY = matrixY;
     }
 
     public <Person> Matrix createMatrix(List<Person> persons, Method... methods) throws InvocationTargetException, IllegalAccessException {
-        int[][] X = new int[persons.size()][methods.length];
-        long[] Y = new long[persons.size()];
+        float[][] X = new float[persons.size()][methods.length];
+        float[] Y = new float[persons.size()];
         for(int i = 0; i < persons.size(); i++){
             Person person = persons.get(i);
             for (int j = 0; j < methods.length; j++) {
                 if(j == 0) {X[i][j] = 0;}
-                else {X[i][j] = (int) methods[j-1].invoke(person);}
+                else {X[i][j] = (float) methods[j-1].invoke(person);}
             }
         }
         for(int i = 0; i < persons.size(); i++){
             Person person = persons.get(i);
-            Y[i] = (long) methods[methods.length-1].invoke(person);
+            Y[i] = (float) methods[methods.length-1].invoke(person);
         }
 
         return new Matrix(X, Y);
     }
 
-    public int[][] getMatrixX() {
+    public float[][] getMatrixX() {
         return matrixX;
     }
 
-    public long[] getMatrixY() {
+    public float[] getMatrixY() {
         return matrixY;
     }
 }
